Refactoring: extracted Kestrel configuration.
